requirements.txt
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.3
Django==5.0.2
django-anymail==10.2
django-environ==0.11.2
djangorestframework==3.14.0
idna==3.6
pycparser==2.21
python-http-client==3.3.7
pytz==2024.1
requests==2.31.0
sendgrid==6.11.0
sqlparse==0.4.4
starkbank-ecdsa==2.2.0
tzdata==2024.1
urllib3==2.2.1
